+++ src/state/settings/mod.rs
@@ -23,7 +23,7 @@ pub struct Settings {
 }
 
 impl Settings {
-    pub fn new(base_time_between_ticks: u16, default_scale: u64) -> Self {
+    pub fn new(base_time_between_ticks: u16, default_scale: u32) -> Self {
         Self {
             speed: Speed::new(base_time_between_ticks),
             scale: Scale::new(default_scale),
@@ -39,7 +39,7 @@ impl Settings {
 }
 
 impl Settings {
-    pub fn reload(&mut self, base_time_between_ticks: u16, default_scale: u64) {
+    pub fn reload(&mut self, base_time_between_ticks: u16, default_scale: u32) {
         self.speed.reload(base_time_between_ticks);
         self.scale.reload(default_scale);
         self.is_running = false;
