+++ src/state/settings/scale.rs
@@ -5,14 +5,14 @@ use gset::Getset;
 pub struct Scale {
     /// Значение
     #[getset(get_copy, vis = "pub")]
-    value: u64,
+    value: u32,
     /// Строковое знчение
     #[getset(get, vis = "pub", ty = "&str")]
     value_string: String,
 }
 
 impl Scale {
-    pub fn new(default_scale: u64) -> Self {
+    pub fn new(default_scale: u32) -> Self {
         Self {
             value: default_scale,
             value_string: default_scale.to_string(),
@@ -21,12 +21,12 @@ impl Scale {
 }
 
 impl Scale {
-    pub fn set(&mut self, scale: u64) {
+    pub fn set(&mut self, scale: u32) {
         self.value = scale;
         self.value_string = scale.to_string();
     }
 
-    pub fn set_value(&mut self, scale: u64) {
+    pub fn set_value(&mut self, scale: u32) {
         if scale > 0 {
             self.value = scale;
         }
@@ -38,7 +38,7 @@ impl Scale {
 }
 
 impl Scale {
-    pub fn reload(&mut self, default_scale: u64) {
+    pub fn reload(&mut self, default_scale: u32) {
         self.value = default_scale;
         self.value_string = default_scale.to_string();
     }
